requirements.txt
allure-python-commons==2.9.44
allure-robotframework==2.9.44

python-dotenv==0.19.2
PyYAML==5.3.1
regex==2021.9.24
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
robotframework==4.1.1
robotframework-appiumlibrary==1.6.3
robotframework-assertion-engine==0.2.0
robotframework-browser==10.0.3
robotframework-faker==5.0.0
robotframework-notifications
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==5.1.3

s3transfer==0.5.0
SecretStorage==2.3.1
selenium==3.141.0
setproctitle==1.1.10
simplejson==3.16.0

urllib3==1.25.8
virtualenv==20.10.0
virtualenv-clone==0.5.7

